LPD-102263 React to feature flag value changes instead of reapplying every delivery
The registrar deferred the FeatureFlagListener service registration to the initial request so that registering it would not replay the current value and sweep every company's segments during startup. Deferring the registration to obtain that property hands the service lifecycle to the component instead of the Service Component Runtime, and it only postpones the problem: the replay still swept on the first request of every boot, re-reading every company's segments to reapply a value the data already held, and redelivery of an unchanged value, whether from a provider restart or a repeated save of the same setting, swept again. The reapplication is also lossy in the enable direction, because reactivating every DEFAULT and REFERRED entry discards the entries an administrator deliberately deactivated.

The services now create governed data matching the flag, and the verify process reconciles data that predates them, so no delivery of an unchanged value has work to do. The listener records the delivered value and rewrites the segments only when a delivery changes it, which makes the registration replay a recording, keeps redeliveries idempotent, and leaves an explicit flip, the one event that carries user intent, as the only trigger of the sweep. The sweep runs inline on the delivering thread, so a failing flip reports its failure to whoever flipped instead of logging it into a queue.

The recorded value lives in the component instance, deliberately not persisted: a change that arrives while the component is not registered to observe it is the verify process's concern, not this listener's, and the first delivery after every registration is a replay of a value the data already matches.

Root cause lineage: ce7d7277d293c created the listener sweeping unconditionally on every delivery, 350f799fb5825 reverted it for processing data during startup, and c51e3893098f5 re-landed it behind the initial request barrier, treating when the replay runs as the problem rather than that the replay reapplies at all.
