--- modules/apps/segments/segments-service/src/main/java/com/liferay/segments/internal/feature/flag/SegmentsFeatureFlagListenerRegistrar.java
@@ -1,146 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.segments.internal.feature.flag;
-
-import com.liferay.petra.function.transform.TransformUtil;
-import com.liferay.portal.kernel.feature.flag.FeatureFlagListener;
-import com.liferay.portal.kernel.model.GroupConstants;
-import com.liferay.portal.kernel.model.GroupModel;
-import com.liferay.portal.kernel.service.CompanyLocalService;
-import com.liferay.portal.kernel.service.GroupLocalService;
-import com.liferay.portal.kernel.servlet.InitialRequestSyncUtil;
-import com.liferay.portal.kernel.util.HashMapDictionaryBuilder;
-import com.liferay.segments.constants.SegmentsEntryConstants;
-import com.liferay.segments.model.SegmentsEntry;
-import com.liferay.segments.model.SegmentsExperience;
-import com.liferay.segments.service.SegmentsEntryLocalService;
-import com.liferay.segments.service.SegmentsExperienceLocalService;
-
-import java.util.List;
-import java.util.Objects;
-
-import org.osgi.framework.BundleContext;
-import org.osgi.framework.ServiceRegistration;
-import org.osgi.service.component.annotations.Activate;
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Deactivate;
-import org.osgi.service.component.annotations.Reference;
-
-/**
- * @author Thiago Buarque
- */
-@Component(service = {})
-public class SegmentsFeatureFlagListenerRegistrar {
-
-	@Activate
-	protected void activate(BundleContext bundleContext) {
-		InitialRequestSyncUtil.registerSyncCallable(
-			() -> {
-				_serviceRegistration = bundleContext.registerService(
-					FeatureFlagListener.class,
-					new SegmentsFeatureFlagListener(),
-					HashMapDictionaryBuilder.put(
-						"feature.flag.key", "LPD-78863"
-					).build());
-
-				return null;
-			});
-	}
-
-	@Deactivate
-	protected void deactivate() {
-		if (_serviceRegistration != null) {
-			_serviceRegistration.unregister();
-		}
-	}
-
-	@Reference
-	private CompanyLocalService _companyLocalService;
-
-	@Reference
-	private GroupLocalService _groupLocalService;
-
-	@Reference
-	private SegmentsEntryLocalService _segmentsEntryLocalService;
-
-	@Reference
-	private SegmentsExperienceLocalService _segmentsExperienceLocalService;
-
-	private ServiceRegistration<FeatureFlagListener> _serviceRegistration;
-
-	private class SegmentsFeatureFlagListener implements FeatureFlagListener {
-
-		@Override
-		public void onValue(
-			long companyId, String featureFlagKey, boolean enabled) {
-
-			if (companyId != 0) {
-				return;
-			}
-
-			_companyLocalService.forEachCompany(
-				company -> {
-					long[] groupIds = TransformUtil.transformToLongArray(
-						_groupLocalService.getGroups(
-							company.getCompanyId(),
-							GroupConstants.ANY_PARENT_GROUP_ID, true),
-						GroupModel::getGroupId);
-
-					if (groupIds.length == 0) {
-						return;
-					}
-
-					_updateSegmentsEntries(enabled, groupIds);
-					_updateSegmentsExperiences(enabled, groupIds);
-				});
-		}
-
-		private void _updateSegmentsEntries(boolean active, long[] groupIds) {
-			List<SegmentsEntry> segmentsEntries =
-				_segmentsEntryLocalService.getSegmentsEntries(
-					groupIds, !active,
-					new String[] {
-						SegmentsEntryConstants.SOURCE_DEFAULT,
-						SegmentsEntryConstants.SOURCE_REFERRED
-					});
-
-			for (SegmentsEntry segmentsEntry : segmentsEntries) {
-				segmentsEntry.setActive(active);
-
-				_segmentsEntryLocalService.updateSegmentsEntry(segmentsEntry);
-			}
-		}
-
-		private void _updateSegmentsExperiences(
-			boolean active, long[] groupIds) {
-
-			List<SegmentsExperience> segmentsExperiences =
-				_segmentsExperienceLocalService.getSegmentsExperiences(
-					groupIds, !active);
-
-			for (SegmentsExperience segmentsExperience : segmentsExperiences) {
-				SegmentsEntry segmentsEntry =
-					_segmentsEntryLocalService.fetchSegmentsEntry(
-						segmentsExperience.getSegmentsEntryId());
-
-				if ((segmentsEntry == null) ||
-					Objects.equals(
-						segmentsEntry.getSource(),
-						SegmentsEntryConstants.SOURCE_ASAH_FARO_BACKEND)) {
-
-					continue;
-				}
-
-				segmentsExperience.setActive(active);
-
-				_segmentsExperienceLocalService.updateSegmentsExperience(
-					segmentsExperience);
-			}
-		}
-
-	}
-
-}
